crossdock - move Wait to crossdock-go
